+++ tomb3decomp/specific/picture.cpp
@@ -6,6 +6,13 @@
 #include "dxshell.h"
 #include "texture.h"
 #include "file.h"
+#ifdef TROYESTUFF
+#include "../tomb3/tomb3.h"
+#endif
+
+#ifdef TROYESTUFF
+bool bDontGreyOut;
+#endif
 
 void CrossFadePicture()
 {
@@ -292,7 +299,20 @@ void CreateMonoScreen()
 {
 	DISPLAYMODE* dm;
 
-	DXTextureSetGreyScale(1);
+#ifdef TROYESTUFF
+	if (bDontGreyOut)
+	{
+		if (tomb3.psx_mono)
+			DXTextureSetGreyScale(0);
+		else
+			DXTextureSetGreyScale(1);
+
+		bDontGreyOut = 0;
+	}
+	else
+#endif
+		DXTextureSetGreyScale(1);
+
 	dm = &App.DeviceInfoPtr->DDInfo[App.DXConfigPtr->nDD].D3DInfo[App.DXConfigPtr->nD3D].DisplayMode[App.DXConfigPtr->nVMode];
 
 	if (dm->w == 640 && dm->h == 480)
@@ -429,9 +449,46 @@ void S_FadeToBlack()
 	TIME_Init();
 }
 
-void DrawMonoScreen(long r, long g, long b)
+void DrawMonoScreen(long r, long g, long b)	//do not call this function with a value higher than 127
 {
-	TRDrawPicture(0, CurPicTexIndices, f_zfar);
+#ifdef TROYESTUFF
+	long x[4];
+	long y[4];
+	static long screenX[4] = { 0, 256, 512, 640 };
+	static long screenY[3] = { 0, 256, 480 };
+	long col;
+
+	if (tomb3.psx_mono)
+	{
+		r <<= 1;
+		g <<= 1;	//compensate for PSX contrast
+		b <<= 1;
+		col = RGBA(r, g, b, 0xFF);
+
+		for (int i = 0; i < 3; i++)
+		{
+			x[i] = phd_winxmin + phd_winwidth * screenX[i] / 640;
+			y[i] = phd_winymin + phd_winheight * screenY[i] / 480;
+		}
+
+		x[3] = phd_winxmin + phd_winwidth * screenX[3] / 640;
+
+		HWR_EnableAlphaBlend(0);
+		HWR_EnablePerspCorrect(0);
+		HWR_EnableColorAddition(0);
+
+		DrawTile(x[0], y[0], x[1] - x[0], y[1] - y[0], CurPicTexIndices[0], 0, 0, 256, 256, col, col, col, col, f_zfar);
+		DrawTile(x[1], y[0], x[2] - x[1], y[1] - y[0], CurPicTexIndices[1], 0, 0, 256, 256, col, col, col, col, f_zfar);
+		DrawTile(x[2], y[0], x[3] - x[2], y[1] - y[0], CurPicTexIndices[2], 0, 0, 128, 256, col, col, col, col, f_zfar);
+		DrawTile(x[0], y[1], x[1] - x[0], y[2] - y[1], CurPicTexIndices[3], 0, 0, 256, 224, col, col, col, col, f_zfar);
+		DrawTile(x[1], y[1], x[2] - x[1], y[2] - y[1], CurPicTexIndices[4], 0, 0, 256, 224, col, col, col, col, f_zfar);
+		DrawTile(x[2], y[1], x[3] - x[2], y[2] - y[1], CurPicTexIndices[2], 128, 0, 128, 224, col, col, col, col, f_zfar);
+
+		HWR_EnablePerspCorrect(1);
+	}
+	else
+#endif
+		TRDrawPicture(0, CurPicTexIndices, f_zfar);
 }
 
 void RemoveMonoScreen(long fade)