+++ tomb3decomp/tomb3/tomb3.cpp
@@ -78,6 +78,7 @@ static void T3_InitSettings()
 	tomb3.upv_wake = 1;
 	tomb3.psx_fov = 0;
 	tomb3.psx_boxes = 0;
+	tomb3.psx_mono = 0;
 	tomb3.shadow_mode = SHADOW_PSX;
 	tomb3.bar_mode = BAR_PSX;
 	tomb3.sophia_rings = SRINGS_PSX;
@@ -143,6 +144,7 @@ void T3_SaveSettings()
 	REG_WriteBool((char*)"upv_wake", tomb3.upv_wake);
 	REG_WriteBool((char*)"psx_fov", tomb3.psx_fov);
 	REG_WriteBool((char*)"psx_boxes", tomb3.psx_boxes);
+	REG_WriteBool((char*)"psx_mono", tomb3.psx_mono);
 	REG_WriteLong((char*)"shadow_mode", tomb3.shadow_mode);
 	REG_WriteLong((char*)"bar_mode", tomb3.bar_mode);
 	REG_WriteLong((char*)"sophia_rings", tomb3.sophia_rings);
@@ -215,6 +217,7 @@ bool T3_LoadSettings()
 	REG_ReadBool((char*)"upv_wake", tomb3.upv_wake, 1);
 	REG_ReadBool((char*)"psx_fov", tomb3.psx_fov, 0);
 	REG_ReadBool((char*)"psx_boxes", tomb3.psx_boxes, 0);
+	REG_ReadBool((char*)"psx_mono", tomb3.psx_mono, 0);
 	REG_ReadLong((char*)"shadow_mode", (ulong&)tomb3.shadow_mode, SHADOW_PSX);
 	REG_ReadLong((char*)"bar_mode", (ulong&)tomb3.bar_mode, BAR_PSX);
 	REG_ReadLong((char*)"sophia_rings", (ulong&)tomb3.sophia_rings, SRINGS_PSX);